Different dialogs for single vs double
