+++ packages/sv-core/src/index.ts
@@ -1,1 +1,2 @@
-export { default as BreakendOptionDialog } from './BreakendOptionDialog'
+export { default as BreakendMultiLevelOptionDialog } from './BreakendMultiLevelOptionDialog'
+export { default as BreakendSingleLevelOptionDialog } from './BreakendSingleLevelOptionDialog'
