+++ plugins/variants/src/VariantFeatureWidget/BreakendMultiLevelOptionDialog.tsx
@@ -1,0 +1,1 @@
+export { BreakendMultiLevelOptionDialog as default } from '@jbrowse/sv-core'
