--- plugins/variants/src/VariantFeatureWidget/BreakendOptionDialog.tsx
@@ -1,5 +1,0 @@
-
-
-
-
-export {BreakendOptionDialog as default} from '@jbrowse/sv-core'
