+++ plugins/variants/src/VariantFeatureWidget/BreakendSingleLevelOptionDialog.tsx
@@ -1,0 +1,1 @@
+export { BreakendSingleLevelOptionDialog as default } from '@jbrowse/sv-core'
